refactor: extract state from App
